third_party/WebKit/LayoutTests/imported/wpt/html/semantics/grouping-content/the-ol-element/ol.start-reflection-2-expected.txt

This is a testharness.js-based test.
FAIL ol.start - reflection test assert_equals: expected 1 but got 0
Harness: the test ran to completion.
